refactor manifest module
