+++ src/plugin/packer/manifest/v1/__tests__/sourcelist.test.ts
