+++ src/plugin/packer/manifest/v1/sourcelist.ts
